+++ CHANGELOG.md
@@ -47,6 +47,7 @@ and this project adheres to
 - vhost: remove upstream linux vhost test kmod code;
 #### Security
 #### Tools
+- util-linux: test 'unshare --bind' with mount-bind.sh;
 - util-linux: add unshare subdir;
 - add iproute, libblockdev, libtracefs;
 #### Template